refactor: Permission Logic
