+++ data/core/src/main/java/com/seugi/data/core/mapper/WorkspacePermissionMapper.kt
@@ -1,7 +1,7 @@
-package com.seugi.data.workspace.mapper
+package com.seugi.data.core.mapper
 
-import com.seugi.data.workspace.model.WorkspacePermissionModel
-import com.seugi.network.workspace.response.WorkspacePermissionResponse
+import com.seugi.data.core.model.WorkspacePermissionModel
+import com.seugi.network.core.response.WorkspacePermissionResponse
 
 fun WorkspacePermissionResponse.toModel() = when (this) {
     WorkspacePermissionResponse.ADMIN -> WorkspacePermissionModel.ADMIN
