+++ data/workspace/src/main/java/com/seugi/data/workspace/WorkspaceRepository.kt
@@ -5,7 +5,7 @@ import com.seugi.data.core.model.ProfileModel
 import com.seugi.data.workspace.model.CheckWorkspaceModel
 import com.seugi.data.workspace.model.WaitWorkspaceModel
 import com.seugi.data.workspace.model.WorkspaceModel
-import com.seugi.data.workspace.model.WorkspacePermissionModel
+import com.seugi.data.core.model.WorkspacePermissionModel
 import kotlinx.coroutines.flow.Flow
 
 interface WorkspaceRepository {
@@ -15,7 +15,6 @@ interface WorkspaceRepository {
 
     suspend fun getMembers(workspaceId: String): Flow<Result<List<ProfileModel>>>
 
-    suspend fun getPermission(workspaceId: String): Flow<Result<WorkspacePermissionModel>>
     suspend fun getMyWorkspaces(): Flow<Result<List<WorkspaceModel>>>
     suspend fun updateWorkspaceId(workspaceId: String)
     suspend fun insertWorkspaceId(workspaceId: String)
