+++ data/workspace/src/main/java/com/seugi/data/workspace/repository/WorkspaceRepositoryImpl.kt
@@ -8,16 +8,19 @@ import com.seugi.data.core.mapper.toModels
 import com.seugi.data.core.model.ProfileModel
 import com.seugi.data.workspace.WorkspaceRepository
 import com.seugi.data.workspace.mapper.localToModel
-import com.seugi.data.workspace.mapper.toModel
+import com.seugi.data.core.mapper.toModel
 import com.seugi.data.workspace.mapper.toModels
 import com.seugi.data.workspace.model.CheckWorkspaceModel
 import com.seugi.data.workspace.model.WaitWorkspaceModel
 import com.seugi.data.workspace.model.WorkspaceModel
-import com.seugi.data.workspace.model.WorkspacePermissionModel
+import com.seugi.data.core.model.WorkspacePermissionModel
+import com.seugi.data.workspace.mapper.toModel
 import com.seugi.local.room.dao.WorkspaceDao
 import com.seugi.local.room.model.WorkspaceEntity
+import com.seugi.network.core.response.ProfileResponse
 import com.seugi.network.core.response.safeResponse
 import com.seugi.network.workspace.WorkspaceDataSource
+import kotlinx.collections.immutable.ImmutableList
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
@@ -53,16 +56,7 @@ class WorkspaceRepositoryImpl @Inject constructor(
 
     override suspend fun getMembers(workspaceId: String): Flow<Result<List<ProfileModel>>> = flow {
         val response = workspaceDatasource.getMembers(workspaceId).safeResponse()
-
         emit(response.toModels())
-    }
-        .flowOn(dispatcher)
-        .asResult()
-
-    override suspend fun getPermission(workspaceId: String): Flow<Result<WorkspacePermissionModel>> = flow {
-        val response = workspaceDatasource.getPermission(workspaceId).safeResponse()
-
-        emit(response.toModel())
     }
         .flowOn(dispatcher)
         .asResult()
