+++ feature-main/notification/src/main/java/com/seugi/notification/NotificationScreen.kt
@@ -52,9 +52,9 @@ import androidx.lifecycle.compose.LifecycleResumeEffect
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.seugi.common.utiles.toTimeString
 import com.seugi.data.notification.model.NotificationModel
-import com.seugi.data.workspace.model.WorkspacePermissionModel
-import com.seugi.data.workspace.model.isAdmin
-import com.seugi.data.workspace.model.isTeacher
+import com.seugi.data.core.model.WorkspacePermissionModel
+import com.seugi.data.core.model.isAdmin
+import com.seugi.data.core.model.isTeacher
 import com.seugi.designsystem.R
 import com.seugi.designsystem.animation.bounceClick
 import com.seugi.designsystem.animation.combinedBounceClick
