+++ feature-main/notification/src/main/java/com/seugi/notification/navigation/NotificationNavigation.kt
@@ -4,7 +4,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.seugi.data.workspace.model.WorkspacePermissionModel
+import com.seugi.data.core.model.WorkspacePermissionModel
 import com.seugi.notification.NotificationScreen
 
 const val NOTIFICATION_ROUTE = "notification"
