+++ network/core/src/main/java/com/seugi/network/core/response/WorkspacePermissionResponse.kt
@@ -1,4 +1,4 @@
-package com.seugi.network.workspace.response
+package com.seugi.network.core.response
 
 enum class WorkspacePermissionResponse {
     ADMIN,
