+++ network/workspace/src/main/java/com/seugi/network/workspace/datasource/WorkspaceDataSourceImpl.kt
@@ -9,7 +9,7 @@ import com.seugi.network.workspace.request.CreateWorkspaceRequest
 import com.seugi.network.workspace.request.WorkspaceApplicationRequest
 import com.seugi.network.workspace.response.CheckWorkspaceResponse
 import com.seugi.network.workspace.response.WaitWorkspaceResponse
-import com.seugi.network.workspace.response.WorkspacePermissionResponse
+import com.seugi.network.core.response.WorkspacePermissionResponse
 import com.seugi.network.workspace.response.WorkspaceResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
